services/auth-service/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
email-validator